requirements.txt
ansicon==1.89.0
APScheduler==3.6.3
arrow==1.2.3
asgiref==3.5.2
attrs==22.1.0
blessed==1.19.1
cachetools==4.2.2
certifi==2022.9.24
Django==4.1
django-debug-toolbar==3.4.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
flake8==5
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
inflection==0.5.1
jinxed==1.2.0
jsonschema==4.17.3
mccabe==0.7.0
pep8-naming==0.13.2
Pillow==9.3.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
tornado==6.1
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
wcwidth==0.2.5
stripe==5.0.0
